Addressed reviewers' comments
- created test_utilities that include eigen vector data checks
- update author names
- provided more details to testing
